Track where related evals in different files (re)define the same subroutine, and ensure that the profiling results are merged for these cases.
This gets somewhat tricky, as we don't learn which lines in which file IDs
need to be merged until we've seen the SUB_INFOs for this profile file, but
those are at the end of the file, after the TIME_LINEs and TIME_BLOCKs.

So read each profile file twice. On the first pass, generate no output, but
store the mapping for file IDs and subroutine info. On the second pass, use
the mappings to modify the other data as appropriate.